score 3

+++ HashCode2020/HashCode2020/Common.h
@@ -30,7 +30,7 @@ struct Library
     vector<unsigned long long> books;
 
     bool is_commited = false;
-    int score = 0;
+    double score = 0;
 };
 
 extern vector<Library> LIBRARIES;
